fix(ci): split provider_readiness part_002 under 1500-line maintainability limit
The coverage-shard maintainability guard
(test_first_party_code_files_stay_under_line_limit) failed because
test_provider_readiness_part_002.py reached 1508 lines after the latest
fixes. Move the trailing opencode/ollama-probe test block into a new
test_provider_readiness_part_009.py with only the imports and helpers it
uses, leaving part_002 at 1267 lines.
